# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Aigoni API
  version: 1.0.0
  description: |
    Aigoni API contract. `x-implementation-status` marks routes registered now
    implemented REST contract.
servers:
  - url: /
...
paths:
...
  /api/content:
    post:
      tags: [Machine API]
      summary: Create a note, post, or page from full Markdown
      description: "Content metadata comes from the YAML frontmatter in `body`. Uses `X-API-Key` or `Authorization: Bearer`; body limit is 2 MiB."
      x-implementation-status: implemented
      security: [{APIKeyHeader: []}, {BearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContentRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MachineContentResponse'
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '502': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
...
  /api/admin/v1/wiki/status:
    get:
      tags: [Admin API]
      summary: Get Wiki configuration and note status counts
      x-implementation-status: implemented
      security: [{AdminSession: []}]
      responses:
        '200':
          description: Wiki status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WikiStatus'
  /api/admin/v1/wiki/chat:
    post:
      tags: [Admin API]
      summary: Start one natural-language Wiki Agent run
      description: Admin Wiki Chat and machine Wiki Ask share the `AIGONI_WIKI_AGENT_CONCURRENCY` active-run pool. Pending and running runs count toward the limit; exceeding it returns 409.
      x-implementation-status: implemented
      security: [{AdminSession: [], CSRFToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WikiChatRequest'}
      responses:
        '202': {description: Agent run accepted, content: {application/json: {schema: {$ref: '#/components/schemas/WikiRun'}}}}
        '409': {$ref: '#/components/responses/Error'}
  /api/admin/v1/wiki/chat:stream:
    post:
      tags: [Admin API]
      summary: Start one Wiki Agent run and stream its events
      description: Response Content-Type is `text/event-stream`. Events are `step`, `tool`, `delta`, `done`, and `error`. The backend run is not tied to the SSE connection lifecycle. Admin Wiki Chat and machine Wiki Ask share the `AIGONI_WIKI_AGENT_CONCURRENCY` active-run pool.
      x-implementation-status: implemented
      security: [{AdminSession: [], CSRFToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WikiChatRequest'}
      responses:
        '200': {description: SSE stream, content: {text/event-stream: {schema: {type: string}}}}
        '409': {$ref: '#/components/responses/Error'}
  /api/admin/v1/wiki/runs/{id}:
    get:
      tags: [Admin API]
      summary: Get a Wiki Agent run
      x-implementation-status: implemented
      security: [{AdminSession: []}]
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses:
        '200': {description: Agent run, content: {application/json: {schema: {$ref: '#/components/schemas/WikiRun'}}}}
        '404': {$ref: '#/components/responses/Error'}
  /api/admin/v1/wiki/runs/{id}/cancel:
    post:
      tags: [Admin API]
      summary: Cancel a running Wiki Agent run
      x-implementation-status: implemented
      security: [{AdminSession: [], CSRFToken: []}]
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses:
        '200': {description: Agent run canceled, content: {application/json: {schema: {$ref: '#/components/schemas/WikiRun'}}}}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /api/admin/v1/wiki/runs/{id}/events:
    get:
      tags: [Admin API]
      summary: Subscribe to Wiki Agent run events
      x-implementation-status: implemented
      security: [{AdminSession: []}]
      parameters:
        - $ref: '#/components/parameters/ID'
        - {name: from, in: query, schema: {type: integer, minimum: 0}}
      responses:
        '200': {description: SSE stream, content: {text/event-stream: {schema: {type: string}}}}
        '404': {$ref: '#/components/responses/Error'}
...
  /api/admin/v1/wiki/documents:
    get:
      tags: [Admin API]
      summary: List Wiki and note documents
      x-implementation-status: implemented
      security: [{AdminSession: []}]
      responses:
        '200':
          description: Documents
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: {$ref: '#/components/schemas/WikiDocument'}
  /api/admin/v1/wiki/documents/content:
    get:
      tags: [Admin API]
      summary: Read and render one allowed Wiki or note document
      x-implementation-status: implemented
      security: [{AdminSession: []}]
      parameters:
        - name: path
          in: query
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Document
          content:
            application/json:
              schema: {$ref: '#/components/schemas/WikiDocument'}
        '404': {$ref: '#/components/responses/Error'}
  /api/admin/v1/wiki/render:
    post:
      tags: [Admin API]
      summary: Render Markdown for Wiki preview
      x-implementation-status: implemented
      security: [{AdminSession: [], CSRFToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [markdown]
              properties: {markdown: {type: string}}
      responses:
        '200':
          description: Rendered HTML
          content:
            application/json:
              schema: {type: object, properties: {html: {type: string}}}
  /api/admin/v1/wiki/backups:
    delete:
      tags: [Admin API]
      summary: Delete Wiki backup files
      x-implementation-status: implemented
      security: [{AdminSession: [], CSRFToken: []}]
      responses:
        '200':
          description: Deleted backup count
          content:
            application/json:
              schema: {type: object, properties: {removed: {type: integer}, message: {type: string}}}
components:
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: aigoni_session
      description: HttpOnly, SameSite=Lax; Secure over HTTPS.
    CSRFToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required on Admin write operations in the target contract.
    APIKeyHeader: {type: apiKey, in: header, name: X-API-Key}
    BearerAuth: {type: http, scheme: bearer}
  parameters:
...
    ID: {name: id, in: path, required: true, schema: {type: string}}
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
  schemas:
...
    CreateContentRequest: {type: object, required: [type, body], properties: {type: {type: string, enum: [note, post, page]}, body: {type: string, description: Full Markdown document with YAML frontmatter. date/lastmod accept YYYY-MM-DD, YYYY-MM-DD HH:MM, YYYY-MM-DD HH:MM:SS, or RFC3339.}, sync_images: {type: boolean, default: false}, publish: {type: boolean, description: Optional. Overrides or supplies frontmatter publish for post/page; note rejects it.}}}
    MachineContentResponse: {type: object, required: [id, path, title, date, synced_images], properties: {id: {type: string}, path: {type: string}, title: {type: string}, date: {type: string, format: date-time}, slug: {type: string}, publish: {type: boolean, description: Present for post and page responses}, toc: {type: boolean, description: Present for post and page responses}, synced_images: {type: integer}}}
...
    WikiStatus: {type: object, properties: {ready: {type: boolean}}}
    WikiRun: {type: object, required: [id, kind, status], properties: {id: {type: string}, kind: {type: string}, status: {type: string, enum: [pending, running, completed, failed]}, message: {type: string}, question: {type: string}, answer: {type: string}, answer_html: {type: string}, reasoning: {type: string}, sources: {type: array, items: {type: string}}, files: {type: array, items: {type: string}}, events: {type: array, items: {$ref: '#/components/schemas/WikiRunEvent'}}, error: {type: string}, created_at: {type: string, format: date-time}, updated_at: {type: string, format: date-time}}}
    WikiRunEvent: {type: object, required: [seq, event, payload], properties: {seq: {type: integer}, time: {type: string, format: date-time}, event: {type: string}, payload: {type: object}}}
    WikiChatRequest:
      type: object
      required: [message]
      properties:
        message: {type: string}
        history:
          type: array
          items:
            type: object
            properties:
              role: {type: string, enum: [user, assistant]}
              content: {type: string}
    WikiDocument: {type: object, required: [path], properties: {path: {type: string}, content: {type: string}, html: {type: string}, meta: {type: object}}}
    ErrorResponse: {type: object, required: [error], properties: {error: {oneOf: [{type: string}, {$ref: '#/components/schemas/ErrorDetail'}]}}}
    ErrorDetail: {type: object, required: [code, message], properties: {code: {type: string}, message: {type: string}, field_errors: {type: object, additionalProperties: {type: string}}, request_id: {type: string}}}